+++ Library/ComplexDataStructures/CustomVector.cs
@@ -6,28 +6,28 @@ namespace GradientDescentSharp.ComplexDataStructures;
 public class CustomVector : Vector
 {
     ///<inheritdoc/>
-    public CustomVector(VectorStorage<double> storage) : base(storage)
+    public CustomVector(VectorStorage<FloatType> storage) : base(storage)
     {
     }
 }
-class CustomVectorStorage : VectorStorage<double>
+class CustomVectorStorage : VectorStorage<FloatType>
 {
     public override bool IsDense => true;
 
     public int StartIndex { get; }
-    public IDataAccess<double> Data { get; }
+    public IDataAccess<FloatType> Data { get; }
 
-    public CustomVectorStorage(IDataAccess<double> data, int startIndex, int length) : base(length)
+    public CustomVectorStorage(IDataAccess<FloatType> data, int startIndex, int length) : base(length)
     {
         StartIndex = startIndex;
         Data = data;
     }
-    public override double At(int index)
+    public override FloatType At(int index)
     {
         return Data[StartIndex + index];
     }
 
-    public override void At(int index, double value)
+    public override void At(int index, FloatType value)
     {
         Data[StartIndex + index] = value;
     }
